COMP: Move Fonts.hpp and LatoRegular.hpp to core simplnx Utilities
* fonts::Base64Decode and the embedded Lato Regular TTF are needed by
  the core ScaleBarRenderer for the Write Image scale bar (#786)
* LatoBold and FiraSansRegular stay in OrientationAnalysis (only used
  by WritePoleFigureFilter)
